requirements.txt
pip>=23.2
PySide6>=6.5.2
Jinja2>=3.1.2
pydantic>=2.3.0
fastapi>=0.103.1
requests>=2.31.0
uvicorn>=0.23.2
packaging>=23.1
psycopg2>=2.9.7
python-multipart>=0.0.6
python-jose>=3.3.0
cryptography>=41.0.3
passlib>=1.7.4
bcrypt>=4.0.1
pyyaml>=6.0.1
